app/requirements.txt
# Requirements for Scopal Boards

# Django
Django==2
django-cms
django-annoying==0.10.3
django-widget-tweaks==1.4.1
icalendar==3.11.7
Markdown==2.6.9
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0

# Frontend
django-webpack-loader

# Rest API
djangorestframework

# Postgres Manager
psycopg2
